ciao-controller, ciao-cli: move volume APIs to ciao API endpoint
Migrate the API routes to the ciao API endpoint and adapt for the ciao
API endpoint versioning. This required removing some definitions and
structures to the api package.

This change also deletes the openstack/block directory as the APIs are
fully migrated.
